Move helpers into JS folder and out of separate spec folder
